+++ src/Renderers/ImageRenderer.php
@@ -5,78 +5,132 @@ namespace BigFish\PDF417\Renderers;
 use BigFish\PDF417\BarcodeData;
 use BigFish\PDF417\RendererInterface;
 
-use Intervention\Image\Image;
+use Intervention\Image\ImageManager;
+use Intervention\Image\Gd\Color;
 
-class ImageRenderer implements RendererInterface
+class ImageRenderer extends AbstractRenderer
 {
     /** Supported image formats and corresponding MIME types. */
-    private $formats = [
+    protected $formats = [
         'jpg' => 'image/jpeg',
         'png' => 'image/png',
         'gif' => 'image/gif',
     ];
 
-    private $options = [
+    protected $options = [
         'format' => 'png',
         'quality' => 90,
         'scale' => 3,
         'ratio' => 3,
         'padding' => 20,
-        'color' => "#000",
-        'bgColor' => "#fff",
+        'color' => "#000000",
+        'bgColor' => "#ffffff",
     ];
 
-    public function __construct(array $options)
+    /**
+     * {@inheritdoc}
+     */
+    public function validateOptions()
     {
-        // Merge given options with defaults
-        foreach ($options as $key => $value) {
-            if (isset($this->options[$key])) {
-                $this->options[$key] = $value;
-            }
-        }
+        $errors = [];
 
-        // Validate options
         $format = $this->options['format'];
         if (!isset($this->formats[$format])) {
-            throw new \InvalidArgumentException("Invalid image format: \"$format\".");
+            $formats = implode(", ", array_keys($this->formats));
+            $errors[] = "Invalid option \"format\": \"$format\". Expected one of: $formats.";
         }
+
+        $scale = $this->options['scale'];
+        if (!is_numeric($scale) || $scale < 1 || $scale > 20) {
+            $errors[] = "Invalid option \"scale\": \"$scale\". Expected an integer between 1 and 20.";
+        }
+
+        $ratio = $this->options['ratio'];
+        if (!is_numeric($ratio) || $ratio < 1 || $ratio > 10) {
+            $errors[] = "Invalid option \"ratio\": \"$ratio\". Expected an integer between 1 and 10.";
+        }
+
+        $padding = $this->options['padding'];
+        if (!is_numeric($padding) || $padding < 0 || $padding > 50) {
+            $errors[] = "Invalid option \"padding\": \"$padding\". Expected an integer between 0 and 50.";
+        }
+
+        $quality = $this->options['quality'];
+        if (!is_numeric($quality) || $quality < 0 || $quality > 100) {
+            $errors[] = "Invalid option \"quality\": \"$quality\". Expected an integer between 0 and 50.";
+        }
+
+        // Check colors by trying to parse them
+        $color = $this->options['color'];
+        $bgColor = $this->options['bgColor'];
+
+        $gdColor = new Color();
+
+        try {
+            $gdColor->parse($color);
+        } catch (\Exception $ex) {
+            $errors[] = "Invalid option \"color\": \"$color\". Supported color formats: \"#000000\", \"rgb(0,0,0)\", or \"rgba(0,0,0,0)\"";
+        }
+
+        try {
+            $gdColor->parse($bgColor);
+        } catch (\Exception $ex) {
+            $errors[] = "Invalid option \"bgColor\": \"$bgColor\". Supported color formats: \"#000000\", \"rgb(0,0,0)\", or \"rgba(0,0,0,0)\"";
+        }
+
+        return $errors;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getContentType()
     {
         $format = $this->options['format'];
         return $this->formats[$format];
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function render(BarcodeData $data)
     {
         $pixelGrid = $data->getPixelGrid();
         $height = count($pixelGrid);
         $width = count($pixelGrid[0]);
 
-        $options = $this->options;
+        // Extract options
+        $bgColor = $this->options['bgColor'];
+        $color = $this->options['color'];
+        $format = $this->options['format'];
+        $padding = $this->options['padding'];
+        $quality = $this->options['quality'];
+        $ratio = $this->options['ratio'];
+        $scale = $this->options['scale'];
 
-        $img = Image::canvas($width, $height, $options['bgColor']);
+        // Create a new image
+        $manager = new ImageManager();
+        $img = $manager->canvas($width, $height, $bgColor);
 
         // Render the barcode
         foreach ($pixelGrid as $y => $row) {
             foreach ($row as $x => $value) {
                 if ($value) {
-                    $img->pixel($options['color'], $x, $y);
+                    $img->pixel($color, $x, $y);
                 }
             }
         }
 
         // Apply scaling & aspect ratio
-        $width *= $options['scale'];
-        $height *= $options['scale'] * $options['ratio'];
+        $width *= $scale;
+        $height *= $scale * $ratio;
         $img->resize($width, $height);
 
         // Add padding
-        $width += 2 * $options['padding'];
-        $height += 2 * $options['padding'];
-        $img->resizeCanvas($width, $height, 'center', false, '#fff');
+        $width += 2 * $padding;
+        $height += 2 * $padding;
+        $img->resizeCanvas($width, $height, 'center', false, $bgColor);
 
-        return $img->encode($options['format']);
+        return $img->encode($format, $quality);
     }
 }