+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 PDF417 Changelog
 ================
 
+0.1.1 (2014-12-26)
+------------------
+
+* Added validation of options in renderers.
+* Fixed a bug in ImageRenderer where padding was not in bgColor, but white.
+* Upgraded Intrevention/Image to v2.
+* Added 'quality' option to ImageRenderer (used for JPG only).
+
 0.1.0 (2014-12-24)
 ------------------